Entrega final

+++ src/main/java/edu/eci/pdsw/sampleprj/dao/ClienteDAO.java
@@ -15,6 +15,6 @@ public interface ClienteDAO {
     
     public void save(Cliente c) throws PersistenceException;
     
-    public Cliente load(int id) throws PersistenceException;
+    public Cliente load(long id) throws PersistenceException;
     
 }
